✓ sound: adds cast
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2489 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ support_gimic/sound/sound.c
@@ -183,7 +183,7 @@ static void filltailsample(UINT nCount)
 	if (nCount)
 	{
 		ptr = sndstream.ptr;
-		nOrgSize = (ptr - sndstream.buffer) / 2;
+		nOrgSize = (UINT)((ptr - sndstream.buffer) / 2);
 		if (nOrgSize == 0)
 		{
 			nSampleL = 0;
